+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/ReductiveRowSpec.java
@@ -14,7 +14,7 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 public class ReductiveRowSpec extends RowSpec {
-    final Field lastFixedField;
+    public final Field lastFixedField;
 
     public ReductiveRowSpec(ProfileFields fields, Map<Field, FieldSpec> fieldToFieldSpec, Field lastFixedField) {
         super(fields, fieldToFieldSpec);
