+++ generator/src/test/java/com/scottlogic/deg/generator/generation/databags/RowSpecDataBagSourceFactoryTests.java
@@ -1,10 +1,14 @@
-package com.scottlogic.deg.generator.fieldspecs;
+package com.scottlogic.deg.generator.generation.databags;
 
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.ProfileFields;
+import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
+import com.scottlogic.deg.generator.fieldspecs.ReductiveRowSpec;
+import com.scottlogic.deg.generator.fieldspecs.RowSpec;
 import com.scottlogic.deg.generator.generation.FieldSpecValueGenerator;
 import com.scottlogic.deg.generator.generation.databags.DataBagSource;
 import com.scottlogic.deg.generator.generation.databags.MultiplexingDataBagSource;
+import com.scottlogic.deg.generator.generation.databags.RowSpecDataBagSourceFactory;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
